Use floats instead of uint15/16
Performance seems to be about the same when operating on native buffers.
4x performance improvement when operating on RGBA float GeglBuffer.

+++ mypaint-tiled-surface.h
@@ -22,8 +22,8 @@ typedef struct {
     int tx;
     int ty;
     gboolean readonly;
-    guint16 *buffer;
-    gpointer context; /* Only to be used by the surface implementations. */
+    float *buffer;
+    gpointer context; /* Only to be used by the surface implemenations. */
     int thread_id;
     int mipmap_level;
 } MyPaintTileRequest;
